Extract keyboard mapping logic into rust-switcher-core crate (#69)
### Motivation
- Isolate the pure RU/EN keyboard-layout mapping logic into a
cross-platform core to begin a gradual extraction of
platform-independent functionality.
- Perform the minimal safe split so the existing Windows app crate keeps
its public module paths and tests unchanged.

### Description
- Turn the repository into a Cargo workspace and add the new library
crate `crates/rust-switcher-core` and a path dependency in the root
crate via `rust-switcher-core = { path = "crates/rust-switcher-core" }`
in `Cargo.toml`.
- Move the mapping implementation into
`rust_switcher_core::text::mapping` and expose the same API
(`ConversionDirection`, `conversion_direction_for_text`,
`convert_ru_en_with_direction`, and `convert_ru_en_bidirectional`).
- Replace the original `src/domain/text/mapping.rs` with a thin
compatibility shim `pub use rust_switcher_core::text::mapping::*;` so
existing imports keep working without rewriting call sites or tests.
- Ensure `convert_ru_en_bidirectional` is a normal `pub fn` in the core
crate (not `#[cfg(test)]`) so dependent crate tests can call it.

### Testing
- Ran `cargo +nightly fmt --check` which completed successfully. 
- Built the project with `cargo +nightly build --features debug-tracing`
and the build succeeded.
- Ran unit tests with `cargo +nightly test --locked` and `cargo test`
and all tests completed successfully.
- Ran `cargo clippy --all-targets -- -D warnings` (and `cargo +nightly
clippy --all-targets --all-features -- -D warnings`) which failed
because of pre-existing `unused` / `dead_code` issues in
`src/input/ring_buffer.rs` and an unused re-export in
`src/domain/text/mod.rs`, not introduced by this split.

------
[Codex
Task](https://chatgpt.com/codex/tasks/task_e_699fab83a8248332b93d16c3a8fff961)
