+++ crates/rust-switcher-core/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "rust-switcher-core"
+version = "0.1.0"
+edition = "2024"
+license = "MIT"
+
+[dependencies]
