+++ crates/rust-switcher-core/src/text/mod.rs
@@ -1,0 +1,1 @@
+pub mod mapping;
